+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.5.6 (TBD)
+
+- Feature: The flag `--preview-url-add-request-headers` can be used to inject key value pair headers in every preview page request
+
 ### 2.5.5 (TBD)
 
 - Change: The traffic-manager now requires permissions to read pods across namespaces even if installed with limited permissions
@@ -10,7 +14,7 @@
 
 - Bugfix: The root daemon will no longer get into a bad state when a disconnect is rapidly followed by a new connect.
 
-- Bugfix: The client will now only watch agents from accessible namespaces, and is also constrained to namespaces  explicitly mapped 
+- Bugfix: The client will now only watch agents from accessible namespaces, and is also constrained to namespaces  explicitly mapped
   using the `connect` command's `--mapped-namespaces` flag.
 
 - Bugfix: The `gather-logs` command will only gather traffic-agent logs from accessible namespaces, and is also constrained to namespaces